input: misc: qti-haptics: 修复空指针、播放时序和死锁问题
check NULL pointer before using it

There might be a rare case that previous play has been terminated and
the next play is generating in VMAX mode so that the play effect pointer
would be set to NULL before the play IRQ being served. So check the
effect pointer before using it in the play IRQ handler.

Change-Id: I2d9ace5d09ffde885b4267e1e78771a169c57bec

Clear PLAY bit after all pattern is played

Clear the PLAY bit only after all pattern bytes are queued to make sure
the whole pattern can be played successfully.

Change-Id: I891e5e46168420c6e4963d67bc77c0f79b2a4a8c

Prevent deadlock by waiting on a negative remaining time.

The contruct on qti_haptics_upload_effect()
    rem = hrtimer_get_remaining(&chip->hap_disable_timer);
    time_us = ktime_to_us(rem);
leads to negative values. If the followed usleep is called with this value
it wait's an uncomfortable range of time...
Guard this call by an check. After that, multiple runs of the
CtsOsTestCases - android.os.cts.VibratorTest ran without issues.

Change-Id: Ife8f55c3b3ceb8bb35d30fe1448a51ab32e0b904

Signed-off-by: Fenglin Wu <fenglinw@codeaurora.org>